scripts/requirements.txt
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
anthropic>=0.90.0
tqdm>=4.66.0
Levenshtein>=0.25.0
